refactor: 코딩 컨벤션 준수 및 그림자 수정
